rm endian
